separated createNewCategory() into several functions.
Budget is planned with going from one big budget to creating sub-budgets. To be able to do that, the treeview creation had to be separated into smalled functions, so that user can add new categories while still not defining any sub-categories (item-type). I also added a button, that now creates a new category.